refactor(test): move views + urls into dojo/test/ui/ [test Phase 5]
Move dojo/test/{views,urls}.py to dojo/test/ui/ and update consumers
(dojo/urls.py include + test_apply_finding_template test). urls.py now imports
views from dojo.test.ui.
