+++ dojo/test/ui/views.py
